Added robots.txt and sitemap.xml files
These files should be added to the /public folder after the build process. We could not find the location to add these files to include under /public directory after the build. Please add it to the appropriate location so that these two files will be included in the /public folder.